backport compilation fix

+++ doc/snippets/MatrixBase_part.cpp
@@ -1,4 +1,4 @@
-Matrix3d m = Matrix3i::Zero();
+Matrix3d m = Matrix3d::Zero();
 m.part<Eigen::StrictlyUpperTriangular>().setOnes();
 cout << "Here is the matrix m:" << endl << m << endl;
 cout << "And let us now compute m*m.adjoint() in a very optimized way" << endl
